v.3 не работает сохранение ролей при редактировании Юзера

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/RoleDao.java
@@ -10,9 +10,9 @@ public interface RoleDao {
 
     List<Role> getAll();
 
-    Role getById(int id);
+    Role getById(Long id);
 
     void update(Role role);
 
-    void delete(int id);
+    void delete(Long id);
 }
