+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -1,7 +1,5 @@
 package ru.kata.spring.boot_security.demo.dao;
 
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import ru.kata.spring.boot_security.demo.model.User;
 
 import java.util.List;
@@ -9,8 +7,8 @@ import java.util.List;
 public interface UserDao {
     void add(User user);
     List<User> getAll();
-    User getById(int id);
+    User getById(Long id);
     void update(User user);
-    void delete(int id);
+    void delete(Long id);
     User findByUsername(String username);
 }
